Reorganize config.functions package into config.nodes.
