--- src/main/java/org/mvplugins/multiverse/core/config/functions/package-info.java
@@ -1,1 +1,0 @@
-package org.mvplugins.multiverse.core.config.functions;
