+++ src/main/java/org/mvplugins/multiverse/core/config/node/ListConfigNode.java
@@ -16,7 +16,12 @@ import io.vavr.control.Try;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import org.mvplugins.multiverse.core.config.functions.*;
+import org.mvplugins.multiverse.core.config.node.functions.DefaultStringParserProvider;
+import org.mvplugins.multiverse.core.config.node.functions.DefaultSuggesterProvider;
+import org.mvplugins.multiverse.core.config.node.functions.NodeStringParser;
+import org.mvplugins.multiverse.core.config.node.functions.NodeSuggester;
+import org.mvplugins.multiverse.core.config.node.serializer.DefaultSerializerProvider;
+import org.mvplugins.multiverse.core.config.node.serializer.NodeSerializer;
 
 /**
  * A config node that contains a list of values.
