+++ src/main/java/org/mvplugins/multiverse/core/config/node/functions/DefaultStringParserProvider.java
@@ -1,4 +1,4 @@
-package org.mvplugins.multiverse.core.config.functions;
+package org.mvplugins.multiverse.core.config.node.functions;
 
 import java.util.HashMap;
 import java.util.Map;
