+++ src/main/java/org/mvplugins/multiverse/core/config/node/functions/NodeSuggester.java
@@ -1,4 +1,4 @@
-package org.mvplugins.multiverse.core.config.functions;
+package org.mvplugins.multiverse.core.config.node.functions;
 
 import java.util.Collection;
