+++ src/main/java/org/mvplugins/multiverse/core/config/node/functions/package-info.java
@@ -1,0 +1,1 @@
+package org.mvplugins.multiverse.core.config.node.functions;
